requirements.txt
google-cloud-core >= 1.3.0
google-cloud-bigquery >= 1.25.0
google-api-core >= 1.16.0
googleapis-common-protos >= 1.6.0
google-auth >= 1.20.0